trying to fix a bug found by Chris: "luban start" not working on mac os x
it turns out that the "-d" option of cherryd does not work if 
"autoreload_on" is set to True. 
In this revision, we don't use "-d" option. that means
users have to use control-D to stop the server.
Need to watch if this plays out well in the long run.